Criação ClienteController e ajustes de layout
