+++ src/main/java/com/alura/easycharge/controller/ClienteController.java
@@ -1,25 +1,13 @@
 package com.alura.easycharge.controller;
 
-import com.alura.easycharge.models.Cliente;
-import com.alura.easycharge.repository.ClienteRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.util.List;
-
-@Controller
+@RequestMapping("cliente")
 public class ClienteController {
 
-    @Autowired
-    private ClienteRepository clienteRepository;
-
-    @GetMapping("/listaCliente")
-    public String lista(Model model){
-
-        List<Cliente> clientes = clienteRepository.findAll();
-        model.addAttribute("clientes", clientes);
-        return "listaCliente";
+    @GetMapping("formulario")
+    public String formulario(){
+        return "cliente/formulario";
     }
 }
