+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.url=jdbc:mysql://localhost:3306/easycharge?useTimezone=true&se
 spring.datasource.username=root
 spring.datasource.password=
 
+
