+++ src/main/resources/templates/cliente/formulario.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Formulário cliente</title>
+</head>
+<body>
+
+</body>
+</html>
