+++ src/main/resources/templates/listaCliente.html
@@ -7,6 +7,19 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 </head>
 <body>
+    <nav class="navbar navbar-expand navbar-light bg-light d-flex justify-content-between mb-3">
+        <!-- <div class="navbar-nav">
+            <a class="nav-item nav-link active" href="">Todos</a>
+            <a class="nav-item nav-link" href="">Aguardando</a>
+            <a class="nav-item nav-link" href="">Aprovado</a>
+            <a class="nav-item nav-link" href="">Entregue</a>
+        </div> -->
+        <a href="cliente/formulario"><button class="btn btn-outline-secondary my-2 my-sm-0">Novo</button></a>
+    </nav>
+
+    <div class="jumbotron mb-0">
+        <h1 class="display-4">Lista de Clientes</h1>
+    </div>
     <table class="table" th:each="cliente : ${clientes}">
         <tr>
             <th scope="col">Nome do Cliente: </th>
